Change the type of the "royalties" variable to "double" instead of "int"

+++ app/src/main/java/com/example/hellocowcow/domain/models/DomainNft.kt
@@ -12,7 +12,7 @@ data class DomainNft (
     var attributes           : String?           = null,
     var name                 : String?           = null,
     var creator              : String?           = null,
-    var royalties            : Int?              = null,
+    var royalties            : Double?              = null,
     var uris                 : ArrayList<String> = arrayListOf(),
     var url                  : String?           = null,
     var thumbnailUrl         : String?           = null,
